Extract Options module
